pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = ""
authors = ["Richard Ramirez <ramirezrichard1709@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.101.1"
python-dotenv = "^1.0.0"
sqlacodegen = "^2.3.0.post1"
psycopg2 = "^2.9.7"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^1.10.7"}
orjson = "^3.8.12"
asyncpg = "^0.27.0"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
pillow = "^10.0.0"
python-multipart = "^0.0.6"
mysqlclient = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
